pyproject.toml
[project]
name = "geradeaus-license-manager-cli"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Thomas Winkel", email = "mail@thomas-winkel.de" }
]
readme = "README.md"
homepage = "https://geradeaus.engineering/"
repository = "https://github.com/ThomasWinkel"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.12"
dependencies = [
    "cryptography>=44.0.0",
    "pyjwt>=2.10.1",
    "typer>=0.15.1",
]

[project.scripts]
geradeaus-lm = "geradeaus_license_manager_cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
